+++ CHANGES.md
@@ -11,6 +11,7 @@
     and [muc_roomid_policy_hint](https://conversejs.org/docs/html/configuration.html#muc-roomid-policy-hint)
 - #1826: A user can now add himself as a contact
 - #1839: Headline messages are shown in controlbox
+- #1834: Add Broadcast-Message (Announcements/MOTD) functionality
 
 ## 6.0.0 (2020-01-09)